move decrypt mnemonic to a new file
