+++ src/views/Shared/SharedView.tsx
@@ -6,7 +6,8 @@ import { ChangeEvent, useEffect, useLayoutEffect, useRef } from 'react';
 import { DriveItemData } from '../../app/drive/types';
 import { useTranslationContext } from '../../app/i18n/provider/TranslationProvider';
 import notificationsService, { ToastType } from '../../app/notifications/services/notifications.service';
-import shareService, { decryptMnemonic } from '../../app/share/services/share.service';
+import shareService from '../../app/share/services/share.service';
+import { decryptMnemonic } from '../../app/share/services/share.crypto';
 import { useAppDispatch, useAppSelector } from '../../app/store/hooks';
 import { storageActions } from '../../app/store/slices/storage';
 import { uiActions } from '../../app/store/slices/ui';
