+++ src/views/Shared/components/SharedItemListContainer.tsx
@@ -12,7 +12,8 @@ import workspacesService from 'services/workspace.service';
 import { OrderDirection } from '../../../app/core/types';
 import { sharedThunks } from '../../../app/store/slices/sharedLinks';
 import workspacesSelectors from '../../../app/store/slices/workspaces/workspaces.selectors';
-import shareService, { decryptMnemonic } from '../../../app/share/services/share.service';
+import shareService from '../../../app/share/services/share.service';
+import { decryptMnemonic } from '../../../app/share/services/share.crypto';
 import { setOrderBy, setPage, setSelectedItems } from '../context/SharedViewContext.actions';
 import { useShareViewContext } from '../context/SharedViewContextProvider';
 import useSharedContextMenu from '../hooks/useSharedContextMenu';
