+++ src/views/Signup/ShareGuestSignUpView.test.tsx
@@ -123,7 +123,6 @@ describe('onSubmit', () => {
         validateSharingInvitation: vi.fn(),
       },
       validateSharingInvitation: vi.fn(),
-      decryptMnemonic: vi.fn(),
     }));
 
     vi.mock('services/navigation.service', () => ({
